+++ src/hyp3_isce2/burst.py
@@ -8,6 +8,7 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Iterator, List, Optional, Tuple, Union
 
+import asf_search
 import isce  # noqa: F401
 import requests
 from isceobj.Sensor.TOPS.Sentinel1 import Sentinel1
@@ -326,25 +327,35 @@ def download_bursts(param_list: Iterator[BurstParams]) -> List[BurstMetadata]:
 def get_product_name(
     reference_scene: str,
     secondary_scene: str,
-    reference_burst_number: int,
-    secondary_burst_number: int,
-    swath_number: int,
-    polarization: str,
 ) -> str:
     """Get the name of the interferogram product.
-    NOTE: Will need to be updated when the interface changes.
 
     Args:
-        reference_scene: The reference scene name.
-        secondary_scene: The secondary scene name.
-        reference_burst_number: The reference burst number.
-        secondary_burst_number: The secondary burst number.
-        swath_number: The swath number.
-        polarization: The polarization.
+        reference_scene: The reference burst name.
+        secondary_scene: The secondary burst name.
 
     Returns:
         The name of the interferogram product.
     """
-    reference_name = f'{reference_scene}_IW{swath_number}_{polarization}_{reference_burst_number}'
-    secondary_name = f'{secondary_scene}_IW{swath_number}_{polarization}_{secondary_burst_number}'
-    return f'{reference_name}x{secondary_name}'
+    return f'{reference_scene}x{secondary_scene}'
+
+
+def search_cmr_uat(scene_name: str) -> asf_search.ASFSearchResults:
+    opts = asf_search.ASFSearchOptions(host='cmr.uat.earthdata.nasa.gov')
+    return asf_search.search(product_list=[scene_name], opts=opts)
+
+
+def get_burst_params(scene_name: str) -> BurstParams:
+    results = search_cmr_uat(scene_name)
+
+    if len(results) == 0:
+        raise ValueError(f'ASF Search failed to find {scene_name}.')
+    if len(results) > 1:
+        raise ValueError(f'ASF Search found multiple results for {scene_name}.')
+
+    return BurstParams(
+        granule=results[0].umm['InputGranules'][0].split('-')[0],
+        swath=results[0].properties['burst']['subswath'],
+        polarization=results[0].properties['polarization'],
+        burst_number=results[0].properties['burst']['burstIndex'],
+    )