+++ src/hyp3_isce2/utils.py
@@ -74,3 +74,9 @@ def make_browse_image(input_tif: str, output_png: str) -> None:
             strict=True,
             scaleParams=[[stats['minimum'], stats['maximum']]],
         )
+
+
+def oldest_granule_first(g1, g2):
+    if g1[14:29] <= g2[14:29]:
+        return g1, g2
+    return g2, g1