+++ tests/test_burst.py
@@ -1,5 +1,7 @@
 from pathlib import Path
+from unittest.mock import patch
 
+import asf_search
 import pytest
 from lxml import etree
 from shapely import geometry
@@ -87,5 +89,70 @@ def test_get_region_of_interest(tmp_path, orbit):
 
 
 def test_get_product_name():
-    name = burst.get_product_name('A', 'B', 1, 2, 3, 'VV')
-    assert name == 'A_IW3_VV_1xB_IW3_VV_2'
+    assert burst.get_product_name('A', 'B') == 'AxB'
+
+
+def mock_asf_search_results(
+        slc_name: str,
+        subswath: str,
+        polarization: str,
+        burst_index: int) -> asf_search.ASFSearchResults:
+    product = asf_search.ASFProduct()
+    product.umm = {'InputGranules': [slc_name]}
+    product.properties.update({
+        'burst': {'subswath': subswath, 'burstIndex': burst_index},
+        'polarization': polarization,
+    })
+    results = asf_search.ASFSearchResults([product])
+    results.searchComplete = True
+    return results
+
+
+def test_get_burst_params_08F8():
+    with patch.object(burst, 'search_cmr_uat') as mock_search_cmr_uat:
+        mock_search_cmr_uat.return_value = mock_asf_search_results(
+            slc_name='S1A_IW_SLC__1SDV_20230526T190821_20230526T190847_048709_05DBA8_08F8-SLC',
+            subswath='IW3',
+            polarization='VV',
+            burst_index=8,
+        )
+        assert burst.get_burst_params('S1_346041_IW3_20230526T190843_VV_08F8-BURST') == burst.BurstParams(
+            granule='S1A_IW_SLC__1SDV_20230526T190821_20230526T190847_048709_05DBA8_08F8',
+            swath='IW3',
+            polarization='VV',
+            burst_number=8,
+        )
+        mock_search_cmr_uat.assert_called_once_with('S1_346041_IW3_20230526T190843_VV_08F8-BURST')
+
+
+def test_get_burst_params_1B3B():
+    with patch.object(burst, 'search_cmr_uat') as mock_search_cmr_uat:
+        mock_search_cmr_uat.return_value = mock_asf_search_results(
+            slc_name='S1A_EW_SLC__1SDH_20230526T143200_20230526T143303_048706_05DB92_1B3B-SLC',
+            subswath='EW5',
+            polarization='HH',
+            burst_index=19,
+        )
+        assert burst.get_burst_params('S1_308695_EW5_20230526T143259_HH_1B3B-BURST') == burst.BurstParams(
+            granule='S1A_EW_SLC__1SDH_20230526T143200_20230526T143303_048706_05DB92_1B3B',
+            swath='EW5',
+            polarization='HH',
+            burst_number=19,
+        )
+        mock_search_cmr_uat.assert_called_with('S1_308695_EW5_20230526T143259_HH_1B3B-BURST')
+
+
+def test_get_burst_params_burst_does_not_exist():
+    with patch.object(burst, 'search_cmr_uat') as mock_search_cmr_uat:
+        mock_search_cmr_uat.return_value = []
+        with pytest.raises(ValueError, match=r'.*failed to find.*'):
+            burst.get_burst_params('this burst does not exist')
+        mock_search_cmr_uat.assert_called_once_with('this burst does not exist')
+
+
+def test_get_burst_params_multiple_results():
+    with patch.object(burst, 'search_cmr_uat') as mock_search_cmr_uat:
+        mock_search_cmr_uat.return_value = ['foo', 'bar']
+        with pytest.raises(ValueError, match=r'.*found multiple results.*'):
+            burst.get_burst_params('there are multiple copies of this burst')
+        mock_search_cmr_uat.assert_called_once_with('there are multiple copies of this burst')